Beginning Steps...
This is still all really primitive and things are likely to change - but
wanted to commit before ending for the night. Decided that panels would
be superior to multiple forms since we would not have to repeat code. To
manage panels, select "View > Other Windows > Document Outline" when in
design view.